For brave coders only (#24)
* Test za routing with data transfer

XD elevator music

* Pain

Prelazak na drugi api

* Formatiranje jsona na transactions

* with error

* Fix errors with transactions

* Delete/Buy/Sell na Frontu

* Dodat overview page

Dodat overview page, funkcionalnost u narednom commitu kada bude bila stabilna

* Name change

Changed the names and paths of some components to match their behavior

* Commit

commit

* Json parsing

parsing

* nesto

* old version/ working / no css/ no api

* auth za delete

* fix za url

+++ backend/api/transactions_api.py
@@ -71,11 +71,17 @@ def create_transaction():
 
 # Obrisi transakciju 
 @transactions_api.route("/api/transactions/<id>", methods=["DELETE"])
+@jwt_required()
 def delete_transaction(id):
+  auth_id = get_jwt_identity()
   try:
     transaction = Transaction.objects.get(id=id)
   except DoesNotExist:
     return not_found()
+  
+  if str(transaction["userid"]) != auth_id:
+    return unauthorized()
+  
 
   transaction.delete()
 
@@ -117,13 +123,6 @@ def get_calculations():
   list_uniq=[] #lista klasa sa koinima koji imaju uniq name
   
 
-  #transaction_type
-  #price
-  #u zavisnosti sta je t type price ide u bought ili sold
-  #def this section as function
-  #assign this function to thread
-  #start it /return list uniq
-  #stop
   for uniqtransaction in users_transactions1:
     if(uniqtransaction["crypto_currency"] not in list_uniq_names):      
       if(uniqtransaction["transaction_type"] == TransactionType.BUY):
@@ -165,8 +164,6 @@ def get_calculations():
           x["amount_sold"] += uniqtransaction["kolicina"]
 
 
-
-      # and uniqtransaction["transaction_type"] == TransactionType.BUY
       
 
   response_data = {
